migrate to a MessageStruct on stack and fix memory alignment issues

+++ control/JobFinishedMessage.cpp
@@ -2,7 +2,7 @@
 #include "JobFinishedMessage.hpp"
 
 
-Synergy::JobFinishedMessage::JobFinishedMessage(char *buffer, int length)
+Synergy::JobFinishedMessage::JobFinishedMessage(char *buffer, size_t length)
     : JobMessage(buffer, length, Message::Type::JobFinished)
 {
